create easy_localization_locale.dart
